requirements.txt
nose==1.3.7
numpy==1.20.1
Pillow==8.1.0
